chore!: new Advz constructor, Advz use u32 instead of usize (#539)
* rename payload_chunk_size -> recovery_threshold, Advz::new -> with_multiplicity, reorder args

* new constructors Advz::new, new_internal with no multiplicity arg

* remove limit on multiplicity arg

* AdvzInternal struct members type usize -> u32 (yikes)

* VidScheme trait functions return u32 instead of usize

* use Advz::new in some tests

* update advz benches with the above changes

* fix rusdoc

+++ primitives/src/vid.rs
@@ -73,13 +73,13 @@ pub trait VidScheme {
     fn is_consistent(commit: &Self::Commit, common: &Self::Common) -> VidResult<()>;
 
     /// Extract the payload byte length data from a [`VidScheme::Common`].
-    fn get_payload_byte_len(common: &Self::Common) -> usize;
+    fn get_payload_byte_len(common: &Self::Common) -> u32;
 
     /// Extract the number of storage nodes from a [`VidScheme::Common`].
-    fn get_num_storage_nodes(common: &Self::Common) -> usize;
+    fn get_num_storage_nodes(common: &Self::Common) -> u32;
 
     /// Extract the number of poly evals per share [`VidScheme::Common`].
-    fn get_multiplicity(common: &Self::Common) -> usize;
+    fn get_multiplicity(common: &Self::Common) -> u32;
 }
 
 /// Convenience struct to aggregate disperse data.
